Denormalize responses for Article and Comment
In order to view details about articles and comments in the application
without making multiple API requests, I have denormalized the data that
is returned with each of these resources by adding user-related and
comment count fields.  This also corrects the issue of not knowing the
user ID of the poster for both resources.